implemented multiple login functionality
