+++ app/Http/Controllers/Auth/LoginController.php
@@ -2,7 +2,12 @@
 
 namespace App\Http\Controllers\Auth;
 
+use App\Admin;
+use \Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
 
@@ -26,7 +31,17 @@ class LoginController extends Controller
      *
      * @var string
      */
-    protected $redirectTo = '/dashboard';
+    // protected $redirectTo = '/dashboard';
+    protected function redirectTo()
+    {
+        if( !empty(Auth::guard('admin')->user()) && (Auth::guard('admin')->user()->admintype == 'superadmin'))
+        {
+            return 'admin/dashboard';
+        }
+        else{
+            return 'home';
+        }
+    }
 
     /**
      * Create a new controller instance.
@@ -36,5 +51,21 @@ class LoginController extends Controller
     public function __construct()
     {
         $this->middleware('guest')->except('logout');
+        $this->middleware('guest:admin')->except('logout');
+    }
+
+    protected function adminLogin(Request $request)
+    {
+        $this->validate($request, [
+            'email'   => 'required|email',
+            'password' => 'required|min:6'
+        ]);
+       
+        if (Auth::guard('admin')->attempt(['email' => $request->email, 'password' => $request->password], $request->get('remember'))) {
+
+
+            return redirect()->intended('/admin/dashboard');
+        }
+        return back()->withInput($request->only('email', 'remember'));
     }
 }
