+++ app/helpers.php
@@ -1,0 +1,8 @@
+<?php
+
+use App\User;
+
+function gender()
+{
+    return User::GENDER;
+}
