+++ database/seeds/AdminSeeder.php
@@ -1,0 +1,25 @@
+<?php
+
+use Carbon\Carbon;
+use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
+
+class AdminSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     *
+     * @return void
+     */
+    public function run()
+    {
+        DB::table('admins')->insert([
+            'name' => 'Bob',
+            'email' => 'bob@gmail.com',
+            'password' => Hash::make('12345678'),
+            'admintype' => 'superadmin',
+            'created_at' => Carbon::now()
+        ]);
+    }
+}
